requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-jose
passlib[bcrypt]
email-validator
python-multipart
python-dotenv
bcrypt==3.2.2


